First version of client binary.
This is a hackathon people.
I have done the dirtiest of dirty and created the client binary in the directory where the library source code resides.
I will rectify this later. The rest of the code needs re-factoring, then I will create a separate fork for go-netconf and a separate repo for the netconf client.

For now, this works. Dirty as hell, but it works. I have a todo in Trello ;)