Close ssh socket even if we get an error
Currently if you get an error when running `t.sshSession.Close()` the
socket will be left open, and there is no way to close it. If this
happens enough times it will cause the device to stop accepting any ssh
connection until the program that has all the open connections is
closed.

Therefore I have updated to close the socket even if we see an error
when running `t.sshSession.Close()`. This way the socket isn't left in
an `ESTABLISHED` state.