transport: major refactor
In order to handle message io.Readers right we need to advance the readers
to the start of the next message (done in framed protocol by reading to
the deliminator).  We do this when the next reader is grabbed.

Similarly if we haven't closed a message writer when we try to get
another message writer it should be an error (we are in the middle of
writing a message) still.

This allows for errors to be propigated correctly to the session so we
can close the `recv` loop properly.  This all makes session closing
actually work.