rewrite framed transports
To make it easier to test and better to use rework the existing
ChunkedTransport and EOMTransport into a single Framer.  Framer
implements MsgReader and MsgWriter like a transport but intentionally
don't have a Close method (and isn't named FramedTransport) as it really
isn't ever intended to be used directly as a transport for a session but
instead by embedded into a transport to prove framing logic.

This makes testing way more straight forward and makes implementing
framed transports (so far SSH and TLS) more straight forward.