hide the "upgrade" interface
The 1.0 to 1.1 upgrade right now is not a stable API and I am unsure
what the stable API should be like.  Since this is only needed for ssh
and tls transport (for now) lets not worry about it and hide the details.

In the future if we want to reexpose it for documentationa and for other
transport to be able to use we can do so.