add capture abilities for underlying framed transports
This adds a method for adding debug capture abilities to capture the
raw output to the underlying writer for framed transports (ssh/tls).

This will aid with development and debugging to see what exactly is
happening.

Also added a GONETCONF_FRAMED_CAPDIR enviroment variable to capture
the raw in/out from the framed transport to capture files.  The idea
is this should make it easy for people to report errors with
reproducable results.