add integration tests for csrx (#8)
Initial support for CI integration tests for running against containerized router control planes in GHA allowing for more validation of changes against really vendor implementations.

First one uses a Juniper cSRX demo image.  This is, unfortunately, not distributuable so it lives as a private image hosted on Github itself.  We'll need to add some instructions for others to be able to replicate this if they want to do local testing, etc. 

Next will be to replicate this for Arista's cEOS  and Nokia SR Linux (which is publicly available.  Go Nokia?)