inttest: add netopeer2 (#16)
Adds support for the opensource project netopeer2. This is the first integration test that can be ran without logins and virtual images from vendors. It is setup for ssh but it can also support tls and call home and should be a very powerful target for integration testing over the vendors.